refactor: replace inline reducers with shared module imports
- Remove inline reducer definitions from all 9 components
- Import reducers and types from src/reducers/index.ts
- Export DetailOutput type from shared reducers module
- Fix pendingDecisionRef type to use PermissionDecision
- Net reduction of ~784 lines of duplicated code
