+++ packages/code/src/components/ConfirmationSelector.tsx
@@ -7,6 +7,11 @@ import {
   ENTER_PLAN_MODE_TOOL_NAME,
   ASK_USER_QUESTION_TOOL_NAME,
 } from "wave-agent-sdk";
+import {
+  confirmationSelectorReducer,
+  questionReducer,
+  type ConfirmationSelectorState,
+} from "../reducers/index.js";
 
 const getHeaderColor = (header: string) => {
   const colors = ["red", "green", "blue", "magenta", "cyan"] as const;
@@ -27,349 +32,6 @@ export interface ConfirmationSelectorProps {
   onCancel: () => void;
 }
 
-interface ConfirmationState {
-  selectedOption: "clear" | "auto" | "allow" | "alternative";
-  alternativeText: string;
-  alternativeCursorPosition: number;
-  hasUserInput: boolean;
-}
-
-type ConfirmationAction =
-  | { type: "SELECT_OPTION"; option: ConfirmationState["selectedOption"] }
-  | { type: "UPDATE_ALTERNATIVE_TEXT"; text: string; cursorPosition: number }
-  | { type: "INPUT_CHARACTER_CONFIRMATION"; input: string }
-  | { type: "DELETE_BEFORE_CURSOR_ALT" }
-  | { type: "MOVE_ALTERNATIVE_CURSOR"; cursorPosition: number }
-  | { type: "SET_HAS_USER_INPUT"; hasUserInput: boolean };
-
-function confirmationReducer(
-  state: ConfirmationState,
-  action: ConfirmationAction,
-): ConfirmationState {
-  switch (action.type) {
-    case "SELECT_OPTION":
-      return { ...state, selectedOption: action.option };
-    case "UPDATE_ALTERNATIVE_TEXT":
-      return {
-        ...state,
-        selectedOption: "alternative",
-        alternativeText: action.text,
-        alternativeCursorPosition: action.cursorPosition,
-        hasUserInput: true,
-      };
-    case "MOVE_ALTERNATIVE_CURSOR": {
-      const delta = action.cursorPosition;
-      const newPos = state.alternativeCursorPosition + delta;
-      return {
-        ...state,
-        alternativeCursorPosition: Math.max(
-          0,
-          Math.min(state.alternativeText.length, newPos),
-        ),
-      };
-    }
-    case "INPUT_CHARACTER_CONFIRMATION": {
-      const text =
-        state.alternativeText.slice(0, state.alternativeCursorPosition) +
-        action.input +
-        state.alternativeText.slice(state.alternativeCursorPosition);
-      return {
-        ...state,
-        selectedOption: "alternative",
-        alternativeText: text,
-        alternativeCursorPosition:
-          state.alternativeCursorPosition + action.input.length,
-        hasUserInput: true,
-      };
-    }
-    case "DELETE_BEFORE_CURSOR_ALT": {
-      if (state.alternativeCursorPosition <= 0) return state;
-      const text =
-        state.alternativeText.slice(0, state.alternativeCursorPosition - 1) +
-        state.alternativeText.slice(state.alternativeCursorPosition);
-      return {
-        ...state,
-        selectedOption: "alternative",
-        alternativeText: text,
-        alternativeCursorPosition: state.alternativeCursorPosition - 1,
-        hasUserInput: text.length > 0,
-      };
-    }
-    case "SET_HAS_USER_INPUT":
-      return { ...state, hasUserInput: action.hasUserInput };
-    default:
-      return state;
-  }
-}
-
-interface QuestionState {
-  currentQuestionIndex: number;
-  selectedOptionIndex: number;
-  selectedOptionIndices: Set<number>;
-  userAnswers: Record<string, string>;
-  otherText: string;
-  otherCursorPosition: number;
-  savedStates: Record<
-    number,
-    {
-      selectedOptionIndex: number;
-      selectedOptionIndices: Set<number>;
-      otherText: string;
-      otherCursorPosition: number;
-    }
-  >;
-}
-
-type QuestionAction =
-  | { type: "SELECT_OPTION_INDEX"; index: number }
-  | { type: "SELECT_OPTION_INDEX_DELTA"; delta: number; maxOptions?: number }
-  | { type: "TOGGLE_CURRENT_OPTION_INDEX" }
-  | { type: "UPDATE_OTHER_TEXT"; text: string; cursorPosition: number }
-  | { type: "APPEND_OTHER_TEXT"; input: string }
-  | { type: "INPUT_CHARACTER"; input: string; optionsCount: number }
-  | { type: "DELETE_BEFORE_CURSOR_OTHER"; maxOptions?: number }
-  | { type: "MOVE_OTHER_CURSOR"; cursorPosition: number; maxOptions?: number }
-  | {
-      type: "NAVIGATE_QUESTION";
-      direction: number;
-      questions: Array<{
-        question: string;
-        options: Array<{ label: string }>;
-        multiSelect?: boolean;
-      }>;
-    }
-  | {
-      type: "CONFIRM_ANSWER";
-      questions: Array<{
-        question: string;
-        options: Array<{ label: string }>;
-        multiSelect?: boolean;
-      }>;
-      pendingDecisionRef: React.MutableRefObject<PermissionDecision | null>;
-    }
-  | { type: "SET_QUESTION_STATE"; state: Partial<QuestionState> };
-
-function questionReducer(
-  state: QuestionState,
-  action: QuestionAction,
-): QuestionState {
-  switch (action.type) {
-    case "SELECT_OPTION_INDEX":
-      return { ...state, selectedOptionIndex: action.index };
-    case "SELECT_OPTION_INDEX_DELTA": {
-      const newIndex = state.selectedOptionIndex + action.delta;
-      const maxIndex = (action.maxOptions ?? Infinity) - 1;
-      return {
-        ...state,
-        selectedOptionIndex: Math.max(0, Math.min(maxIndex, newIndex)),
-      };
-    }
-    case "TOGGLE_CURRENT_OPTION_INDEX": {
-      const nextIndices = new Set(state.selectedOptionIndices);
-      if (nextIndices.has(state.selectedOptionIndex)) {
-        nextIndices.delete(state.selectedOptionIndex);
-      } else {
-        nextIndices.add(state.selectedOptionIndex);
-      }
-      return { ...state, selectedOptionIndices: nextIndices };
-    }
-    case "UPDATE_OTHER_TEXT":
-      return {
-        ...state,
-        otherText: action.text,
-        otherCursorPosition: action.cursorPosition,
-      };
-    case "INPUT_CHARACTER": {
-      const otherIdx = (action.optionsCount || 0) - 1;
-      if (state.selectedOptionIndex !== otherIdx) return state;
-      const text =
-        state.otherText.slice(0, state.otherCursorPosition) +
-        action.input +
-        state.otherText.slice(state.otherCursorPosition);
-      return {
-        ...state,
-        otherText: text,
-        otherCursorPosition: state.otherCursorPosition + action.input.length,
-      };
-    }
-    case "APPEND_OTHER_TEXT": {
-      const text =
-        state.otherText.slice(0, state.otherCursorPosition) +
-        action.input +
-        state.otherText.slice(state.otherCursorPosition);
-      return {
-        ...state,
-        otherText: text,
-        otherCursorPosition: state.otherCursorPosition + action.input.length,
-      };
-    }
-    case "DELETE_BEFORE_CURSOR_OTHER": {
-      const otherIdx = (action.maxOptions ?? 0) - 1;
-      if (state.selectedOptionIndex !== otherIdx) return state;
-      if (state.otherCursorPosition <= 0) return state;
-      return {
-        ...state,
-        otherText:
-          state.otherText.slice(0, state.otherCursorPosition - 1) +
-          state.otherText.slice(state.otherCursorPosition),
-        otherCursorPosition: state.otherCursorPosition - 1,
-      };
-    }
-    case "MOVE_OTHER_CURSOR": {
-      const otherIdx = (action.maxOptions ?? 0) - 1;
-      if (state.selectedOptionIndex !== otherIdx) return state;
-      const delta = action.cursorPosition;
-      const newPos = state.otherCursorPosition + delta;
-      return {
-        ...state,
-        otherCursorPosition: Math.max(
-          0,
-          Math.min(state.otherText.length, newPos),
-        ),
-      };
-    }
-    case "NAVIGATE_QUESTION": {
-      const questions = action.questions;
-      let nextIndex = state.currentQuestionIndex + action.direction;
-      if (nextIndex < 0) nextIndex = questions.length - 1;
-      if (nextIndex >= questions.length) nextIndex = 0;
-
-      if (nextIndex === state.currentQuestionIndex) return state;
-
-      const savedStates = {
-        ...state.savedStates,
-        [state.currentQuestionIndex]: {
-          selectedOptionIndex: state.selectedOptionIndex,
-          selectedOptionIndices: state.selectedOptionIndices,
-          otherText: state.otherText,
-          otherCursorPosition: state.otherCursorPosition,
-        },
-      };
-
-      const nextState = savedStates[nextIndex] || {
-        selectedOptionIndex: 0,
-        selectedOptionIndices: new Set<number>(),
-        otherText: "",
-        otherCursorPosition: 0,
-      };
-
-      return {
-        ...state,
-        currentQuestionIndex: nextIndex,
-        ...nextState,
-        savedStates,
-      };
-    }
-    case "CONFIRM_ANSWER": {
-      const questions = action.questions;
-      const currentQuestion = questions[state.currentQuestionIndex];
-      if (!currentQuestion) return state;
-
-      const options = [...currentQuestion.options, { label: "Other" }];
-      const isOtherFocused = state.selectedOptionIndex === options.length - 1;
-      let answer = "";
-      if (currentQuestion.multiSelect) {
-        const selectedLabels = Array.from(state.selectedOptionIndices)
-          .filter((i) => i < currentQuestion.options.length)
-          .map((i) => currentQuestion.options[i].label);
-        const isOtherChecked = state.selectedOptionIndices.has(
-          options.length - 1,
-        );
-        if (isOtherChecked && state.otherText.trim()) {
-          selectedLabels.push(state.otherText.trim());
-        }
-        answer = selectedLabels.join(", ");
-      } else {
-        if (isOtherFocused) {
-          answer = state.otherText.trim();
-        } else {
-          answer = options[state.selectedOptionIndex].label;
-        }
-      }
-
-      if (!answer) return state;
-
-      const newAnswers = {
-        ...state.userAnswers,
-        [currentQuestion.question]: answer,
-      };
-
-      if (state.currentQuestionIndex < questions.length - 1) {
-        const nextIndex = state.currentQuestionIndex + 1;
-        const savedStates = {
-          ...state.savedStates,
-          [state.currentQuestionIndex]: {
-            selectedOptionIndex: state.selectedOptionIndex,
-            selectedOptionIndices: state.selectedOptionIndices,
-            otherText: state.otherText,
-            otherCursorPosition: state.otherCursorPosition,
-          },
-        };
-
-        const nextState = savedStates[nextIndex] || {
-          selectedOptionIndex: 0,
-          selectedOptionIndices: new Set<number>(),
-          otherText: "",
-          otherCursorPosition: 0,
-        };
-
-        return {
-          ...state,
-          currentQuestionIndex: nextIndex,
-          ...nextState,
-          userAnswers: newAnswers,
-          savedStates,
-        };
-      } else {
-        const finalAnswers = { ...newAnswers };
-        for (const [idxStr, s] of Object.entries(state.savedStates)) {
-          const idx = parseInt(idxStr);
-          const q = questions[idx];
-          if (q && !finalAnswers[q.question]) {
-            const opts = [...q.options, { label: "Other" }];
-            let a = "";
-            if (q.multiSelect) {
-              const selectedLabels = Array.from(s.selectedOptionIndices)
-                .filter((i) => i < q.options.length)
-                .map((i) => q.options[i].label);
-              const isOtherChecked = s.selectedOptionIndices.has(
-                opts.length - 1,
-              );
-              if (isOtherChecked && s.otherText.trim()) {
-                selectedLabels.push(s.otherText.trim());
-              }
-              a = selectedLabels.join(", ");
-            } else {
-              if (s.selectedOptionIndex === opts.length - 1) {
-                a = s.otherText.trim();
-              } else {
-                a = opts[s.selectedOptionIndex].label;
-              }
-            }
-            if (a) finalAnswers[q.question] = a;
-          }
-        }
-
-        const allAnswered = questions.every((q) => finalAnswers[q.question]);
-        if (!allAnswered) return state;
-
-        action.pendingDecisionRef.current = {
-          behavior: "allow",
-          message: JSON.stringify(finalAnswers),
-        };
-        return {
-          ...state,
-          userAnswers: finalAnswers,
-        };
-      }
-    }
-    case "SET_QUESTION_STATE":
-      return { ...state, ...action.state };
-    default:
-      return state;
-  }
-}
-
 export const ConfirmationSelector: React.FC<ConfirmationSelectorProps> = ({
   toolName,
   toolInput,
@@ -379,7 +41,7 @@ export const ConfirmationSelector: React.FC<ConfirmationSelectorProps> = ({
   onDecision,
   onCancel,
 }) => {
-  const [state, dispatch] = useReducer(confirmationReducer, {
+  const [state, dispatch] = useReducer(confirmationSelectorReducer, {
     selectedOption: toolName === EXIT_PLAN_MODE_TOOL_NAME ? "clear" : "allow",
     alternativeText: "",
     alternativeCursorPosition: 0,
@@ -590,7 +252,7 @@ export const ConfirmationSelector: React.FC<ConfirmationSelectorProps> = ({
       }
     }
 
-    const availableOptions: ConfirmationState["selectedOption"][] = [];
+    const availableOptions: ConfirmationSelectorState["selectedOption"][] = [];
     if (toolName === EXIT_PLAN_MODE_TOOL_NAME) availableOptions.push("clear");
     availableOptions.push("allow");
     if (!hidePersistentOption) availableOptions.push("auto");
